Fix compiler warnings.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@19868 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/SVN/SVNPatch.cpp
@@ -153,7 +153,7 @@ int SVNPatch::Init( const CString& patchfile, const CString& targetpath, CProgre
         return 0;
     }
     
-    if ((m_nRejected > (m_filePaths.size() / 3)) && !m_testPath.IsEmpty())
+    if ((m_nRejected > ((int)m_filePaths.size() / 3)) && !m_testPath.IsEmpty())
     {
         m_nStrip++;
         bool found = false;
